feat(install): add package info to json output (#8234)
This was removed in npm 7 and never got re-added. In order for this not
to be a breaking change the info was added to new fields, since the old
fields are already being used to return counts.

This differs slightly from the npm 6 output, as "changed" now has a
"from" and "to" field instead of a tacked on "previousVersion" field.

Closes: https://github.com/npm/cli/issues/6558